S3-JS Renamed project to MonoZelda (#16)
* Renamed project to MonoZelda

* Consistant JobName so we can Branch reulset check on it.
