--- LevelBasedGame/Program.cs
@@ -1,3 +1,0 @@
-﻿
-using var game = new LevelBasedGame.LevelGame();
-game.Run();
