+++ MonoZelda.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35208.52
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LevelBasedGame", "LevelBasedGame.csproj", "{677E556A-0777-4995-A905-EC6D74EC9912}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MonoZelda", "MonoZelda/MonoZelda.csproj", "{677E556A-0777-4995-A905-EC6D74EC9912}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
