+++ MonoZelda/Commands/ResetCommand.cs
@@ -1,17 +1,19 @@
 ﻿using System;
+using PixelPushers.MonoZelda.Controllers;
 
-namespace LevelBasedGame.Commands;
-public class ExitCommand : ICommand
+namespace PixelPushers.MonoZelda.Commands;
+
+public class ResetCommand : ICommand
 {
     IController controller;
-    public ExitCommand(IController controller)
+    public ResetCommand(IController controller)
     {
         this.controller = controller;
     }
 
     public GameState Execute()
     {
-        return GameState.Quit;
+        return GameState.Reset;
     }
 
     public GameState UnExecute()
