+++ MonoZelda/Content/Sprite Source Rects - Player.csv
